+++ AirCasting/AirCastingApp.swift
@@ -14,9 +14,7 @@ struct AirCastingApp: App {
     var body: some Scene {
             WindowGroup {
                 NavigationView {
-//                    GraphView()
-//                    Dashboard()
-                    SessionCell()
+                    MainTabBarView()
 
             }
         }
